defense/orchestrator/requirements.txt
fastapi
uvicorn
python-dotenv
rich
httpx
websockets

